requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
Werkzeug==3.0.1
python-dotenv==1.0.1
requests==2.31.0
logging==0.4.9.6 